# Major refactoring of CLI code and interface