--- a/data/input_file.txt
+++ b/data/input_file.txt
@@ -1,0 +1,2 @@
+data/test
+data/recursive/1/images