… data/recursive/1/images/cam_order.npy
… data/recursive/1/images/camera_0_img_000000.jpg
… data/recursive/1/images/camera_0_img_000001.jpg
… data/recursive/1/images/camera_0_img_000002.jpg
… data/recursive/1/images/camera_0_img_000003.jpg
… data/recursive/1/images/camera_0_img_000004.jpg
… data/recursive/1/images/camera_0_img_000005.jpg
… data/recursive/1/images/camera_0_img_000006.jpg
… data/recursive/1/images/camera_0_img_000007.jpg
… data/recursive/1/images/camera_0_img_000008.jpg
… data/recursive/1/images/camera_0_img_000009.jpg
… data/recursive/1/images/camera_0_img_000010.jpg
… data/recursive/1/images/camera_0_img_000011.jpg
… data/recursive/1/images/camera_0_img_000012.jpg
… data/recursive/1/images/camera_0_img_000013.jpg
… data/recursive/1/images/camera_0_img_000014.jpg
… data/recursive/1/images/camera_1_img_000000.jpg
… data/recursive/1/images/camera_1_img_000001.jpg
… data/recursive/1/images/camera_1_img_000002.jpg
… data/recursive/1/images/camera_1_img_000003.jpg
… data/recursive/1/images/camera_1_img_000004.jpg
… data/recursive/1/images/camera_1_img_000005.jpg
… data/recursive/1/images/camera_1_img_000006.jpg
… data/recursive/1/images/camera_1_img_000007.jpg
… data/recursive/1/images/camera_1_img_000008.jpg
… data/recursive/1/images/camera_1_img_000009.jpg
… data/recursive/1/images/camera_1_img_000010.jpg
… data/recursive/1/images/camera_1_img_000011.jpg
… data/recursive/1/images/camera_1_img_000012.jpg
… data/recursive/1/images/camera_1_img_000013.jpg