**This pull request:**
1. implements the `--from-file` flag to read a list of folders from a text file ;
2. implements a major refactoring of the CLI interface (better argument names) ;
3. implements new flags to choose the verbosity (`--verbose` and `--verbose2`) ;
4. is a major refactoring of the code: the new files `core_api.py` and `utils.py` were created ;
5. adds better error handling in `pose2d/drosophilia.py` and fixes some deprecated warnings.
6. adds bold text in output for easier reading :-)

More details below. You can also check the git commits, I kept them "clean".

## Code refactoring

**(1)** I have replaced the `print` statements by `logger.debug` instead, this allows to choose the verbosity level through command line arguments `-v` or `-vv`.

**(2)** All the functions that should have been implemented in the core of deepfly instead of the CLI interface are now in `CLI/core_api.py`, and the code exclusively related to the CLI is in `CLI/main.py`. This makes a better decoupling from the CLI and the core and makes the CLI easier to read.

**(3)** I also attempted to "die with more meaningful errors" in case something unexpected happens. So I renamed a few error message and added more error detection. For instance, instead of having an exception saying:

```
Error: max of empty list
```

The code now says:

```
Error: No image found.
```


## New command line interface

```
(deepfly) user@upramdyapc6:~/julien/DeepFly3D$ df3d-cli -h

usage: df3d-cli [-h] [-v] [-vv] [-d] [-r] [-f] [-n NUM_IMAGES_MAX]
                [-i [CAMERA_IDS [CAMERA_IDS ...]]] [-2d] [-3d] [-skip]
                INPUT

DeepFly3D pose estimation

positional arguments:
  INPUT                 Without additional arguments, a folder containing
                        unlabeled images.

optional arguments:
  -h, --help            show this help message and exit
  -v, --verbose         Enable info output (such as progress bars)
  -vv, --verbose2       Enable debug output
  -d, --debug           Displays the argument list for debugging purposes
  -r, --recursive       INPUT is a folder. Successively use its subfolders
                        named 'images/'
  -f, --from-file       INPUT is a text-file, where each line names a folder.
                        Successively use the listed folders.
  -n NUM_IMAGES_MAX, --num-images-max NUM_IMAGES_MAX
                        Maximal number of images to process.
  -i [CAMERA_IDS [CAMERA_IDS ...]], --camera-ids [CAMERA_IDS [CAMERA_IDS ...]]
                        Ordering of the cameras provided as an ordered list of
                        ids. Example: 0 1 4 3 2 5 6.
  -2d, --video-2d       Generate pose2d videos
  -3d, --video-3d       Generate pose3d videos
  -skip, --skip-estimation
                        Skip pose estimation
```

**Examples:**

- Simple folder + 3d videos: `df3d-cli -v -3d data/test`
- From file + 3d videos........: `df3d-cli -v -3d -f data/input_file.txt`
- Recursive + 3d videos.......: `df3d-cli -v -3d -r data/recursive/`


## Difference between verbosity levels:

- When there is no verbosity flag, the code outputs only errors. This makes it easier to check if everything went well.
- With the flag `-v` or `--verbose`, the code outputs info on the progress: some progress bars and which folder is being processed.
- With the flag `-vv` or `--verbose2`, the code outputs all debug info.

Here are some demonstrations:

**Without any verbosity flag:** no output
```
(deepfly) user@upramdyapc6:~/julien/DeepFly3D$ df3d-cli -3d -f data/input_file.txt -v
```
**With the verbosity flag `-v`:** see what is being done
```
(deepfly) user@upramdyapc6:~/julien/DeepFly3D$ df3d-cli -3d -f data/input_file.txt -v
Looking for folders listed in data/input_file.txt
Folders found:
-data/test
-data/recursive/1/images

Working in data/test
Processing |################################| (2/2) D:0.022717s|B:0.669s| ...
Video created at /home/user/julien/DeepFly3D/data/test/df3d/pose3d.mp4

Working in data/recursive/1/images
Processing |################################| (2/2) D:0.022598s|B:0.671s| ...
Video created at /home/user/julien/DeepFly3D/data/recursive/1/images/df3d/pose3d.mp4
```

**With the verbosity flag `-vv`:** see all debug output

```
(deepfly) user@upramdyapc6:~/julien/DeepFly3D$ df3d-cli -3d -f data/input_file.txt -vv
Looking for folders listed in data/input_file.txt
Folders found:
-data/test
-data/recursive/1/images
Working in data/test
Creating model 'hg', stacks=8, blocks=1
Loading checkpoint '/home/user/julien/DeepFly3D/deepfly/GUI/../../weights/sh8_deepfly.tar'
Loading the optimizer
Setting image resolution and heatmap resolution: [256, 512] [64, 128]
Loaded checkpoint '/home/user/julien/DeepFly3D/deepfly/GUI/../../weights/sh8_deepfly.tar' (epoch 106)
Total params: 25.61M
Going to process 15 images
Searching for manual corrections
Folders inside validation data: {'/home/user/julien/DeepFly3D/data/test'}
Successfully imported 90 Images in Drosophila Dataset
Predictions shape (8, 15, 19, 2)
Processing |################################| (2/2) D:0.022883s|B:0.667s|L:0.00000000|Acc: 0.0000|Mse: 42.501|F-T: 79.055|T-Tar: 72.044|Tar-tip: 60.714
val_score_maps have shape: (8, 15, 19, 64, 128)
Saving Results, flipping heatmaps
Flipping heatmaps for images with cam_id: [4, 5, 6]
Finished saving results
Looking for data in /home/user/julien/DeepFly3D/data/test/df3d/
/home/user/julien/DeepFly3D/data/test, ['/home/user/julien/DeepFly3D/data/test/df3d/preds_-home-user-julien-DeepFly3D-data-test.pkl']
Loading predictions ['/home/user/julien/DeepFly3D/data/test/df3d/preds_-home-user-julien-DeepFly3D-data-test.pkl']
Loading heatmaps ['/home/user/julien/DeepFly3D/data/test/df3d/heatmap_-home-user-julien-DeepFly3D-data-test.pkl']
Heatmap shape: (8, 15, 19, 64, 128)
Ignore_list [0, 1, 5, 6, 10, 11, 15, 19, 20, 24, 25, 29, 30, 34]:  1.4783
Replaced 0 points
Number of points: 180
/home/user/.conda/envs/deepfly/lib/python3.6/site-packages/scipy/sparse/lil.py:518: FutureWarning: future versions will not create a writeable array from broadcast_array. Set the writable flag explicitly to avoid this warning.
  if not x.flags.writeable:
   Iteration     Total nfev        Cost      Cost reduction    Step norm     Optimality   
       0              1         2.0165e+03                                    4.12e+03    
       1              2         1.9310e+03      8.55e+01       1.35e+00       1.60e+03    
       2              3         1.9282e+03      2.85e+00       3.16e-02       2.16e+01    
       3              4         1.9282e+03      2.29e-03       5.66e-03       1.14e+00    
`ftol` termination condition is satisfied.
Function evaluations 4, initial cost 2.0165e+03, final cost 1.9282e+03, first-order optimality 1.14e+00.
Bundle adjustment, Average reprojection error: 1.445942763378291
Ignore_list [0, 1, 5, 6, 10, 11, 15, 19, 20, 24, 25, 29, 30, 34]:  1.7986
Replaced 0 points
Number of points: 180
   Iteration     Total nfev        Cost      Cost reduction    Step norm     Optimality   
       0              1         3.2312e+03                                    6.21e+03    
       1              2         3.1701e+03      6.11e+01       5.23e-01       5.40e+02    
       2              3         3.1695e+03      6.25e-01       4.25e-02       4.72e+01    
       3              4         3.1695e+03      6.86e-03       5.32e-03       4.14e+00    
`ftol` termination condition is satisfied.
Function evaluations 4, initial cost 3.2312e+03, final cost 3.1695e+03, first-order optimality 4.14e+00.
Bundle adjustment, Average reprojection error: 1.761072189906793
Saving video to: /home/user/julien/DeepFly3D/data/test/df3d/pose3d.mp4
Video size is: (500, 333)
15it [00:01,  7.54it/s]
Video created at /home/user/julien/DeepFly3D/data/test/df3d/pose3d.mp4

Working in data/recursive/1/images
Creating model 'hg', stacks=8, blocks=1
Loading checkpoint '/home/user/julien/DeepFly3D/deepfly/GUI/../../weights/sh8_deepfly.tar'
Loading the optimizer
Setting image resolution and heatmap resolution: [256, 512] [64, 128]
Loaded checkpoint '/home/user/julien/DeepFly3D/deepfly/GUI/../../weights/sh8_deepfly.tar' (epoch 106)
Total params: 25.61M
Going to process 15 images
Searching for manual corrections
Folders inside validation data: {'/home/user/julien/DeepFly3D/data/recursive/1/images'}
Successfully imported 90 Images in Drosophila Dataset
Predictions shape (8, 15, 19, 2)
Processing |################################| (2/2) D:0.022443s|B:0.681s|L:0.00000000|Acc: 0.0000|Mse: 42.501|F-T: 79.055|T-Tar: 72.044|Tar-tip: 60.714
val_score_maps have shape: (8, 15, 19, 64, 128)
Saving Results, flipping heatmaps
Flipping heatmaps for images with cam_id: [4, 5, 6]
Finished saving results
Looking for data in /home/user/julien/DeepFly3D/data/recursive/1/images/df3d/
/home/user/julien/DeepFly3D/data/recursive/1/images, ['/home/user/julien/DeepFly3D/data/recursive/1/images/df3d/preds_-home-user-julien-DeepFly3D-data-recursive-1-images.pkl']
Loading predictions ['/home/user/julien/DeepFly3D/data/recursive/1/images/df3d/preds_-home-user-julien-DeepFly3D-data-recursive-1-images.pkl']
Loading heatmaps ['/home/user/julien/DeepFly3D/data/recursive/1/images/df3d/heatmap_-home-user-julien-DeepFly3D-data-recursive-1-images.pkl']
Heatmap shape: (8, 15, 19, 64, 128)
Ignore_list [0, 1, 5, 6, 10, 11, 15, 19, 20, 24, 25, 29, 30, 34]:  1.4783
Replaced 0 points
Number of points: 180
/home/user/.conda/envs/deepfly/lib/python3.6/site-packages/scipy/sparse/lil.py:518: FutureWarning: future versions will not create a writeable array from broadcast_array. Set the writable flag explicitly to avoid this warning.
  if not x.flags.writeable:
   Iteration     Total nfev        Cost      Cost reduction    Step norm     Optimality   
       0              1         2.0165e+03                                    4.12e+03    
       1              2         1.9310e+03      8.55e+01       1.35e+00       1.60e+03    
       2              3         1.9282e+03      2.85e+00       3.16e-02       2.16e+01    
       3              4         1.9282e+03      2.29e-03       5.66e-03       1.14e+00    
`ftol` termination condition is satisfied.
Function evaluations 4, initial cost 2.0165e+03, final cost 1.9282e+03, first-order optimality 1.14e+00.
Bundle adjustment, Average reprojection error: 1.445942763378291
Ignore_list [0, 1, 5, 6, 10, 11, 15, 19, 20, 24, 25, 29, 30, 34]:  1.7986
Replaced 0 points
Number of points: 180
   Iteration     Total nfev        Cost      Cost reduction    Step norm     Optimality   
       0              1         3.2312e+03                                    6.21e+03    
       1              2         3.1701e+03      6.11e+01       5.23e-01       5.40e+02    
       2              3         3.1695e+03      6.25e-01       4.25e-02       4.72e+01    
       3              4         3.1695e+03      6.86e-03       5.32e-03       4.14e+00    
`ftol` termination condition is satisfied.
Function evaluations 4, initial cost 3.2312e+03, final cost 3.1695e+03, first-order optimality 4.14e+00.
Bundle adjustment, Average reprojection error: 1.761072189906793
Saving video to: /home/user/julien/DeepFly3D/data/recursive/1/images/df3d/pose3d.mp4
Video size is: (500, 333)
15it [00:01,  7.56it/s]
Video created at /home/user/julien/DeepFly3D/data/recursive/1/images/df3d/pose3d.mp4

```